Refactored newWallet

+++ packages/keymaster/src/db-wallet-json.js
@@ -1,18 +1,19 @@
 import fs from 'fs';
-import * as cipher from '@macterra/cipher';
-import * as exceptions from '@macterra/exceptions';
 
 const dataFolder = 'data';
 const walletName = `${dataFolder}/wallet.json`;
 
-export function saveWallet(wallet) {
-    // TBD validate wallet before saving
+export function saveWallet(wallet, overwrite = false) {
+    if (fs.existsSync(walletName) && !overwrite) {
+        return false;
+    }
 
     if (!fs.existsSync(dataFolder)) {
         fs.mkdirSync(dataFolder, { recursive: true });
     }
 
     fs.writeFileSync(walletName, JSON.stringify(wallet, null, 4));
+    return true;
 }
 
 export function loadWallet() {
@@ -21,35 +22,5 @@ export function loadWallet() {
         return JSON.parse(walletJson);
     }
 
-    return newWallet();
-}
-
-export function newWallet(mnemonic, overwrite) {
-    if (fs.existsSync(walletName) && !overwrite) {
-        throw new Error(exceptions.UPDATE_FAILED);
-    }
-
-    try {
-        if (!mnemonic) {
-            mnemonic = cipher.generateMnemonic();
-        }
-        const hdkey = cipher.generateHDKey(mnemonic);
-        const keypair = cipher.generateJwk(hdkey.privateKey);
-        const backup = cipher.encryptMessage(keypair.publicJwk, keypair.privateJwk, mnemonic);
-
-        const wallet = {
-            seed: {
-                mnemonic: backup,
-                hdkey: hdkey.toJSON(),
-            },
-            counter: 0,
-            ids: {},
-        }
-
-        saveWallet(wallet);
-        return wallet;
-    }
-    catch (error) {
-        throw new Error(exceptions.INVALID_PARAMETER);
-    }
+    return null;
 }
